# docs/openapi/mnemonic-v1.yaml
openapi: 3.1.0
info:
  title: Mnemonic Admin REST API
  description: |
    Admin REST API for Mnemonic, a knowledge graph and tooling synchronization server.
    This API manages patterns, agents, and skills. It is consumed by curl or any HTTP
    client for data ingestion and administration.

    The MCP server (port 8081) is a separate protocol for Claude Code integration and
    is not documented in this OpenAPI specification.

    ## Architecture

    Mnemonic is a single Go binary that exposes two protocols:

    - **Admin REST API** (this spec) on port 8080: Read + Write via HTTP
    - **MCP Server** on port 8081: Read-only tools for Claude Code via MCP over Streamable HTTP

    ## Authentication

    MVP: no authentication. The server runs on localhost in a trusted environment.

    Post-MVP: authentication and authorization handled externally by Envoy and OPA.

    ## Error Handling

    All errors follow [RFC 7807 Problem Details](https://tools.ietf.org/html/rfc7807)
    with content type `application/problem+json`. Every response includes an
    `X-Request-ID` correlation header.

    ## Pagination

    List endpoints use cursor-based pagination. Cursors are opaque base64-encoded
    strings. The `has_more` field indicates whether additional pages exist.

    ## Request Correlation

    Every response includes an `X-Request-ID` header. If the client provides this
    header on the request, the server echoes it back. Otherwise the server generates
    a UUID v4.
...
servers:
...
components:
  parameters:
    agentNameParam:
      name: name
      in: path
      required: true
      description: Agent unique identifier (lowercase letters, numbers, and single hyphens; no leading/trailing hyphens, no consecutive hyphens)
      schema:
        type: string
        pattern: ^[a-z]([a-z0-9](-[a-z0-9])*)*$
        maxLength: 64
      example: go-software-engineer
...
    cursorParam:
      name: cursor
      in: query
      description: Opaque pagination cursor from a previous response
      schema:
        type: string
  headers:
    X-Request-ID:
      description: Correlation ID for request tracing. Echoed from request or generated by server.
      schema:
        type: string
        format: uuid
      example: 550e8400-e29b-41d4-a716-446655440000
    Location:
      description: URI of the newly created resource
      schema:
        type: string
        format: uri-reference
  schemas:
...
    ErrorResponse:
      type: object
      description: RFC 7807 Problem Details error response
      required:
        - type
        - title
        - status
        - traceId
      properties:
        type:
          type: string
          format: uri
          description: URI identifying the problem type
          examples:
            - https://mnemonic.example.com/problems/validation-error
            - https://mnemonic.example.com/problems/not-found
        title:
          type: string
          description: Human-readable summary of the problem type
          examples:
            - Validation Error
            - Not Found
        status:
          type: integer
          description: HTTP status code
          minimum: 400
          maximum: 599
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence
        instance:
          type: string
          format: uri-reference
          description: URI reference identifying the specific occurrence
        traceId:
          type: string
          format: uuid
          description: Correlation ID for debugging and log correlation
        errors:
          type: array
          description: Field-level validation errors (present on 400 responses)
          items:
            $ref: "#/components/schemas/FieldError"
    FieldError:
      type: object
      description: Individual field validation error
      required:
        - field
        - code
        - message
      properties:
        field:
          type: string
          description: JSON path to the field with the error
          examples:
            - name
            - system_prompt
        code:
          type: string
          description: Machine-readable error code
          enum:
            - REQUIRED
            - INVALID_FORMAT
            - TOO_LONG
            - TOO_SHORT
            - INVALID_VALUE
            - CONTENT_TOO_LARGE
            - NOT_UNIQUE
        message:
          type: string
          description: Human-readable error message

    # ==================== Pagination Schema ====================
    Pagination:
      type: object
      description: Cursor-based pagination metadata
      required:
        - limit
        - has_more
      properties:
        limit:
          type: integer
          description: Items per page (echoed from request)
        cursor:
          type:
            - string
            - "null"
          description: Current cursor (echoed from request, null for first page)
        next_cursor:
          type:
            - string
            - "null"
          description: Cursor for next page (null when no more pages exist)
        has_more:
          type: boolean
          description: Whether more pages exist after this one

    # ==================== Agent Schemas ====================
    Agent:
      type: object
      description: |
        Full agent definition. Wire format uses flat fields. The server marshals
        flat fields to/from a JSONB `definition` column internally. The `id` and
        `crc64` fields are server-generated and read-only.
      required:
        - id
        - name
        - description
        - system_prompt
        - model
        - version
        - crc64
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Server-generated UUID primary key
          readOnly: true
          examples:
            - 550e8400-e29b-41d4-a716-446655440010
        name:
          type: string
          description: |
            Unique identifier. Lowercase letters, numbers, and single hyphens.
            No leading/trailing hyphens, no consecutive hyphens.
          pattern: ^[a-z]([a-z0-9](-[a-z0-9])*)*$
          minLength: 1
          maxLength: 64
          examples:
            - go-software-engineer
        description:
          type: string
          description: Human-readable description of the agent's purpose
          minLength: 1
          maxLength: 500
          examples:
            - "Implements Go code: functions, packages, tests"
        system_prompt:
          type: string
          description: Full system prompt text (up to 50KB)
          minLength: 1
          maxLength: 51200
        model:
          type: string
          description: Claude model for the agent
          enum:
            - sonnet
            - opus
            - haiku
        allowed_tools:
          type: array
          description: Tool names the agent is allowed to use
          maxItems: 50
          items:
            type: string
          default: []
          examples:
            - - Read
              - Write
              - Edit
              - Bash
              - Glob
              - Grep
        version:
          type: string
          description: Version of the agent definition
          minLength: 1
          maxLength: 50
          examples:
            - "1.2.0"
        crc64:
          type: string
          description: CRC-64 checksum of the serialized definition (server-computed, decimal string)
          readOnly: true
          examples:
            - "1234567890123456789"
        created_at:
          type: string
          format: date-time
          description: Creation timestamp (ISO 8601)
          readOnly: true
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp (ISO 8601)
          readOnly: true
    AgentCreate:
      type: object
      description: Request body for creating a new agent
      required:
        - name
        - description
        - system_prompt
        - model
        - version
      properties:
        name:
          type: string
          pattern: ^[a-z]([a-z0-9](-[a-z0-9])*)*$
          minLength: 1
          maxLength: 64
          examples:
            - rust-software-engineer
        description:
          type: string
          minLength: 1
          maxLength: 500
          examples:
            - "Implements Rust code: crates, modules, tests"
        system_prompt:
          type: string
          minLength: 1
          maxLength: 51200
        model:
          type: string
          enum:
            - sonnet
            - opus
            - haiku
        allowed_tools:
          type: array
          maxItems: 50
          items:
            type: string
          default: []
        version:
          type: string
          minLength: 1
          maxLength: 50
          examples:
            - "1.0.0"
    AgentUpdate:
      type: object
      description: Request body for updating an agent (full replacement). Name in body must match path parameter or be omitted.
      required:
        - description
        - system_prompt
        - model
        - version
      properties:
        name:
          type: string
          description: Must match the path parameter (or be omitted)
          pattern: ^[a-z]([a-z0-9](-[a-z0-9])*)*$
          minLength: 1
          maxLength: 64
        description:
          type: string
          minLength: 1
          maxLength: 500
        system_prompt:
          type: string
          minLength: 1
          maxLength: 51200
        model:
          type: string
          enum:
            - sonnet
            - opus
            - haiku
        allowed_tools:
          type: array
          maxItems: 50
          items:
            type: string
        version:
          type: string
          minLength: 1
          maxLength: 50
    AgentList:
      type: object
      description: Paginated list of agents
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Agent"
        pagination:
          $ref: "#/components/schemas/Pagination"

    # ==================== Pattern Schemas ====================
...
  responses:
    BadRequest:
      description: Validation error or malformed request
      headers:
        X-Request-ID:
          $ref: "#/components/headers/X-Request-ID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            type: https://mnemonic.example.com/problems/validation-error
            title: Validation Error
            status: 400
            detail: The request body contains invalid fields
            instance: /v1/api/agents
            traceId: 550e8400-e29b-41d4-a716-446655440000
            errors:
              - field: name
                code: INVALID_FORMAT
                message: "Name must match pattern ^[a-z]([a-z0-9](-[a-z0-9])*)*$"
              - field: system_prompt
                code: REQUIRED
                message: System prompt is required
    NotFound:
      description: Resource not found
      headers:
        X-Request-ID:
          $ref: "#/components/headers/X-Request-ID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            type: https://mnemonic.example.com/problems/not-found
            title: Not Found
            status: 404
            detail: "Agent with name 'unknown-agent' not found"
            instance: /v1/api/agents/unknown-agent
            traceId: 550e8400-e29b-41d4-a716-446655440000
    Conflict:
      description: Resource conflict (e.g., duplicate name)
      headers:
        X-Request-ID:
          $ref: "#/components/headers/X-Request-ID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            type: https://mnemonic.example.com/problems/conflict
            title: Conflict
            status: 409
            detail: "Agent with name 'go-software-engineer' already exists"
            traceId: 550e8400-e29b-41d4-a716-446655440000
...
    InternalError:
      description: Internal server error
      headers:
        X-Request-ID:
          $ref: "#/components/headers/X-Request-ID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            type: https://mnemonic.example.com/problems/internal-error
            title: Internal Error
            status: 500
            detail: An unexpected error occurred
            traceId: 550e8400-e29b-41d4-a716-446655440000
    ServiceUnavailable:
      description: A required dependency (PostgreSQL, PGVector, or Neo4j) is unavailable
      headers:
        X-Request-ID:
          $ref: "#/components/headers/X-Request-ID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            type: https://mnemonic.example.com/problems/service-unavailable
            title: Service Unavailable
            status: 503
            detail: PGVector is not reachable
            traceId: 550e8400-e29b-41d4-a716-446655440000
paths:
...
  /v1/api/agents:
    get:
      tags:
        - Agents
      summary: List agents
      description: |
        Returns a paginated list of all agents with full definitions including
        system prompts. The default page size is 100 since the collection is small
        (tens of agents).
      operationId: listAgents
      parameters:
        - name: limit
          in: query
          description: Maximum items per page
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 100
        - $ref: "#/components/parameters/cursorParam"
      responses:
        "200":
          description: Paginated list of agents
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentList"
              example:
                data:
                  - id: 550e8400-e29b-41d4-a716-446655440010
                    name: go-software-engineer
                    description: "Implements Go code: functions, packages, tests"
                    system_prompt: "You are an expert Go engineer..."
                    model: sonnet
                    allowed_tools:
                      - Read
                      - Write
                      - Edit
                      - Bash
                      - Glob
                      - Grep
                    version: "1.2.0"
                    crc64: "1234567890123456789"
                    created_at: "2026-01-15T10:30:00Z"
                    updated_at: "2026-02-10T14:00:00Z"
                pagination:
                  limit: 100
                  cursor: null
                  next_cursor: null
                  has_more: false
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags:
        - Agents
      summary: Create agent
      description: |
        Creates a new agent definition. The agent name must be unique and follow
        the pattern: lowercase letters, numbers, and hyphens, starting with a letter.
      operationId: createAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentCreate"
            example:
              name: rust-software-engineer
              description: "Implements Rust code: crates, modules, tests"
              system_prompt: "You are an expert Rust engineer..."
              model: sonnet
              allowed_tools:
                - Read
                - Write
                - Edit
                - Bash
              version: "1.0.0"
      responses:
        "201":
          description: Agent created
          headers:
            Location:
              $ref: "#/components/headers/Location"
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/api/agents/{name}:
    parameters:
      - $ref: "#/components/parameters/agentNameParam"

    get:
      tags:
        - Agents
      summary: Get agent by name
      description: Returns the full agent definition including system prompt.
      operationId: getAgent
      responses:
        "200":
          description: Agent details
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
              example:
                id: 550e8400-e29b-41d4-a716-446655440010
                name: go-software-engineer
                description: "Implements Go code: functions, packages, tests"
                system_prompt: "You are an expert Go engineer..."
                model: sonnet
                allowed_tools:
                  - Read
                  - Write
                  - Edit
                  - Bash
                  - Glob
                  - Grep
                version: "1.2.0"
                crc64: "1234567890123456789"
                created_at: "2026-01-15T10:30:00Z"
                updated_at: "2026-02-10T14:00:00Z"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      tags:
        - Agents
      summary: Update agent
      description: |
        Full replacement of an agent definition. The name in the request body
        must match the path parameter (or be omitted).
      operationId: updateAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentUpdate"
      responses:
        "204":
          description: No Content
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags:
        - Agents
      summary: Delete agent
      description: |
        Deletes an agent. Associated pattern_agent_associations rows are deleted
        via ON DELETE CASCADE.
      operationId: deleteAgent
      responses:
        "204":
          description: Agent deleted
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ==================== Pattern Endpoints ====================
...